requirements.txt
PyYAML==3.11
lazr.uri==1.0.3
wadllib==1.3.2
wsgiref==0.1.2
